Refactor: Modularize rendering pipeline into separate files
