+++ BlenderProc/reflection/resources/splits/abo/split_test.txt
@@ -1,0 +1,20 @@
+B082XLC9B4
+B07QNYVDH4
+B08GGSFH94
+B07DMLG8Y4
+B07B4GQ964
+B07JM1ZP24
+B075X61W74
+B07JK86YC4
+B07SRSJW74
+B0824F6HZ4
+B07QFP4ZG4
+B07P8STD74
+B075X4VW74
+B07MF1SRG4
+B07B7B6W64
+B07TR68JS4
+B071J7Q3J4
+B075X1TB14
+B00I0VTOO4
+B07B4Z9Q14
